Add refactored user.rb, blog_post.rb, model.rb.
